* Converted some ints to smallints (shorts) for basic type tables * Changed primary key to Languages table to Language ISO Code. * Made a few other minor schema tweaks * Rescaffolded views and rebaselined migrations

+++ MovieDbLite.MVC/Models/Award.cs
@@ -10,7 +10,7 @@ namespace MovieDbLite.MVC.Models
         }
 
         public int Id { get; set; }
-        public int AwardShowId { get; set; }
+        public short AwardShowId { get; set; }
         public string AwardName { get; set; }
         public string Description { get; set; }
