+++ MovieDbLite.MVC/Models/AwardShowInstance.cs
@@ -5,7 +5,7 @@ namespace MovieDbLite.MVC.Models
     public partial class AwardShowInstance
     {
         public int Id { get; set; }
-        public int AwardShowId { get; set; }
+        public short AwardShowId { get; set; }
         public string Year { get; set; }
         public DateTime DateHosted { get; set; }
