+++ MovieDbLite.MVC/Models/User.cs
@@ -11,7 +11,7 @@ namespace MovieDbLite.MVC.Models
         }
 
         public int Id { get; set; }
-        public int UserRoleId { get; set; }
+        public short UserRoleId { get; set; }
         public string UserName { get; set; }
         public string EmailAddress { get; set; }
         public string HashedPassword { get; set; }
